post-merge fixes (to build generic 8 bit phy)

+++ include/dev/bb_i2c.h
@@ -24,7 +24,7 @@ void bb_i2c_start(struct i2c_bus *bus);
 void bb_i2c_repeat_start(struct i2c_bus *bus);
 void bb_i2c_stop(struct i2c_bus *bus);
 void bb_i2c_get_byte(struct i2c_bus *bus, uint8_t *data, int last);
-uint8_t bb_i2c_put_byte(struct i2c_bus *bus, uint8_t data);
+int bb_i2c_put_byte(struct i2c_bus *bus, uint8_t data);
 void bb_i2c_delay(uint32_t delay);
 void bb_i2c_scan(struct i2c_bus *bus);
